feat: support ga telemetry (#2761)
* refactor: safary telemetry

* chore: support ga

* fix: logs

* fix: typo
