+++ src/helpers/runInSafe.ts
@@ -2,8 +2,8 @@ export function runInSafe<T>(fn: () => T, noThrow = true) {
     try {
         return fn();
     } catch (error) {
-        if (noThrow) console.error(`[runInSafe] ${error}`);
-        else throw error;
+        if (!noThrow) throw error;
+        console.error(`[runInSafe] ${error}`);
         return;
     }
 }
@@ -12,8 +12,8 @@ export async function runInSafeAsync<T>(fn: () => Promise<T>, noThrow = true) {
     try {
         return await fn();
     } catch (error) {
-        if (noThrow) console.error(`[runInSafeAsync] ${error}`);
-        else throw error;
+        if (!noThrow) throw error;
+        console.error(`[runInSafeAsync] ${error}`);
         return;
     }
 }
