+++ src/providers/telemetry/captureAccountEvent.ts
@@ -5,7 +5,7 @@ import { UnreachableError } from '@/constants/error.js';
 import { createLookupTableResolver } from '@/helpers/createLookupTableResolver.js';
 import { getProfileState } from '@/helpers/getProfileState.js';
 import { runInSafe } from '@/helpers/runInSafe.js';
-import { SafaryTelemetryProvider } from '@/providers/safary/Telemetry.js';
+import { TelemetryProvider } from '@/providers/telemetry/index.js';
 import type { Account } from '@/providers/types/Account.js';
 import { EventId } from '@/providers/types/Telemetry.js';
 
@@ -68,19 +68,19 @@ function getAccountEventParameters(account: Account) {
 export function captureAccountLoginEvent(account: Account) {
     runInSafe(() => {
         const source = account.profile.source;
-        SafaryTelemetryProvider.captureEvent(resolveLoginEventId(source), getAccountEventParameters(account));
+        TelemetryProvider.captureEvent(resolveLoginEventId(source), getAccountEventParameters(account));
     });
 }
 
 export function captureAccountLogoutEvent(account: Account) {
     runInSafe(() => {
         const source = account.profile.source;
-        SafaryTelemetryProvider.captureEvent(resolveLogoutEventId(source), getAccountEventParameters(account));
+        TelemetryProvider.captureEvent(resolveLogoutEventId(source), getAccountEventParameters(account));
     });
 }
 
 export function captureAccountLogoutAllEvent() {
     runInSafe(() => {
-        SafaryTelemetryProvider.captureEvent(EventId.ACCOUNT_LOG_OUT_ALL_SUCCESS, {});
+        TelemetryProvider.captureEvent(EventId.ACCOUNT_LOG_OUT_ALL_SUCCESS, {});
     });
 }
