+++ src/providers/telemetry/captureComposeEvent.ts
@@ -2,9 +2,9 @@ import { safeUnreachable } from '@masknet/kit';
 
 import { UnreachableError } from '@/constants/error.js';
 import { runInSafe } from '@/helpers/runInSafe.js';
-import { getComposeEventParameters, type Options } from '@/providers/safary/getComposeEventParameters.js';
-import { getPostEventId, getPostEventParameters } from '@/providers/safary/getPostEventParameters.js';
-import { SafaryTelemetryProvider } from '@/providers/safary/Telemetry.js';
+import { getComposeEventParameters, type Options } from '@/providers/telemetry/getComposeEventParameters.js';
+import { getPostEventId, getPostEventParameters } from '@/providers/telemetry/getPostEventParameters.js';
+import { TelemetryProvider } from '@/providers/telemetry/index.js';
 import { EventId } from '@/providers/types/Telemetry.js';
 import type { CompositePost } from '@/store/useComposeStore.js';
 import type { ComposeType } from '@/types/compose.js';
@@ -17,7 +17,7 @@ export function captureComposeEvent(type: ComposeType, post: CompositePost, opti
 
         // draft created
         if (draftId) {
-            SafaryTelemetryProvider.captureEvent(EventId.COMPOSE_DRAFT_CREATE_SUCCESS, {
+            TelemetryProvider.captureEvent(EventId.COMPOSE_DRAFT_CREATE_SUCCESS, {
                 draft_id: draftId,
                 draft_time: date.getTime(),
                 draft_time_utc: date.toUTCString(),
@@ -27,7 +27,7 @@ export function captureComposeEvent(type: ComposeType, post: CompositePost, opti
 
         // scheduled post created
         if (scheduleId) {
-            SafaryTelemetryProvider.captureEvent(EventId.COMPOSE_SCHEDULED_POST_CREATE_SUCCESS, {
+            TelemetryProvider.captureEvent(EventId.COMPOSE_SCHEDULED_POST_CREATE_SUCCESS, {
                 schedule_id: scheduleId,
                 schedule_time: date.getTime(),
                 scheduled_time_utc: date.toUTCString(),
@@ -41,7 +41,7 @@ export function captureComposeEvent(type: ComposeType, post: CompositePost, opti
             if (size === 1) {
                 switch (type) {
                     case 'compose':
-                        SafaryTelemetryProvider.captureEvent(
+                        TelemetryProvider.captureEvent(
                             getPostEventId(type, post),
                             getComposeEventParameters(post, options),
                         );
@@ -56,7 +56,7 @@ export function captureComposeEvent(type: ComposeType, post: CompositePost, opti
                         const postId = post.parentPost[source]?.postId;
                         if (!postId) throw new Error(`Target post ID is missing, source = ${source}.`);
 
-                        SafaryTelemetryProvider.captureEvent(
+                        TelemetryProvider.captureEvent(
                             getPostEventId(type, post),
                             getPostEventParameters(postId, profile),
                         );
@@ -68,7 +68,7 @@ export function captureComposeEvent(type: ComposeType, post: CompositePost, opti
 
                 // crossed post
             } else if (size > 1) {
-                SafaryTelemetryProvider.captureEvent(
+                TelemetryProvider.captureEvent(
                     EventId.COMPOSE_CROSS_POST_SEND_SUCCESS,
                     getComposeEventParameters(post, options),
                 );
