+++ src/providers/telemetry/capturePostActionEvent.ts
@@ -2,8 +2,8 @@ import { type SocialSource, Source } from '@/constants/enum.js';
 import { UnreachableError } from '@/constants/error.js';
 import { createLookupTableResolver } from '@/helpers/createLookupTableResolver.js';
 import { runInSafe } from '@/helpers/runInSafe.js';
-import { getPostEventParameters } from '@/providers/safary/getPostEventParameters.js';
-import { SafaryTelemetryProvider } from '@/providers/safary/Telemetry.js';
+import { getPostEventParameters } from '@/providers/telemetry/getPostEventParameters.js';
+import { TelemetryProvider } from '@/providers/telemetry/index.js';
 import type { Post } from '@/providers/types/SocialMedia.js';
 import { EventId } from '@/providers/types/Telemetry.js';
 
@@ -68,6 +68,6 @@ export function capturePostActionEvent(action: PostActionType, post: Post) {
         const eventIds = resolvePostActionEventIds(post.source);
         const eventId = eventIds[action];
 
-        SafaryTelemetryProvider.captureEvent(eventId, getPostEventParameters(post.postId, post.author));
+        TelemetryProvider.captureEvent(eventId, getPostEventParameters(post.postId, post.author));
     });
 }
