+++ src/providers/telemetry/getComposeEventParameters.ts
