+++ src/providers/telemetry/getEventParameters.ts
