+++ src/services/crossSchedulePostThread.ts
@@ -14,7 +14,7 @@ import type { SchedulePayload } from '@/helpers/resolveCreateSchedulePostPayload
 import { EnableSignlessModalRef } from '@/modals/controls.js';
 import { fireflySessionHolder } from '@/providers/firefly/SessionHolder.js';
 import { FireflySocialMediaProvider } from '@/providers/firefly/SocialMedia.js';
-import { captureComposeEvent } from '@/providers/safary/captureComposeEvent.js';
+import { captureComposeEvent } from '@/providers/telemetry/captureComposeEvent.js';
 import { createSchedulePostsPayload } from '@/services/crossSchedulePost.js';
 import { uploadSessions } from '@/services/metrics.js';
 import { useComposeStateStore } from '@/store/useComposeStore.js';
